Okay, everything has been moved to the event structure now
